Remade structure of the network to simplify controls.
Sink is now a full-scale task agent and I fully incorporated
endpoint transmitter after the sink.

Also, during rewriting foud a possible bug in JADE when
trying to use dynamic number of agent parameters
with the help of an ArrayList casted into an Array.
It's no use since JADE's last update was around 2017, but. . .